refactor(catalog): keep filters with usePluginsByFilter hook
